fix: restructure groups and elements
