+++ hermes/crates/cbork/abnf-parser/tests/testsuite/main.rs
@@ -4,6 +4,7 @@ mod alternations;
 mod character_sets;
 mod comments;
 mod elements;
+mod groups;
 mod identifiers;
 mod repetitions;
 mod rules;
